# Conflicts:
#	src/Graph.cpp
#	src/Graph.h
#	src/GraphManager.cpp
#	src/GraphManager.h
#	src/VertexEdge.h

+++ src/VertexEdge.cpp
@@ -39,7 +39,7 @@ int Vertex::getId() const {
 string Vertex::getName() const {
     return this->name;
 }
-int Vertex::getDist() const {
+double Vertex::getDist() const {
     return this->dist;
 }
 double Vertex::getLatitude() const {
@@ -67,7 +67,7 @@ void Vertex::setId(int id) {
 void Vertex::setName(string name) {
     this->name = name;
 }
-void Vertex::setDist(int dist) {
+void Vertex::setDist(double dist) {
     this->dist = dist;
 }
 void Vertex::setLatitude(double lat) {
